Updated Lab 1 Report
The Lab 1 Report file somehow ended up in the Recycle Bin on my
computer; however, the file was the only candidate available on my
computer for the final report. After briefly checking it out, I believe
that the file (which is being committed) is the final version that was
turned in during class.

Note that this file includes the portions of the report which I composed
as well as portions that I rewrote to equalize voice and style across
the report.